tiny_socket: fix around socket.socket.readline(), handling EINTR
Use of httplib inside python.twistd framework has revealed a shortcoming
of the socket.socket.readline(): it would not handle EINTR, interrupted
reads or even other resumable errors correctly. All contents of the read
buffer were popped, meaning that any further attempt to resume reading
from the socket would have lost some bytes.
With this patch, we allow readline to use the read buffer and tolerate
EINTR.
Note that read() is not patched accordingly, because so far, in httplib,
it is used in non-waiting context. It is readline() which blocks more.

Then, override HTTPResponse, so that it installs the hacked readline()
class.