extra: 'options' module for generic stock configuration
It has become a pattern, that many client scripts actually need the same
kind of command-line parameters, logging setup and eventually a config
file to read from.
For all these cases that we don't have strict-strange requirements on
these options, have this stock module to take care of them.

Also include a sample, that performs a task in ~30 LoC