utils: idexed resources for Pool (aka. dictionary mode)
Until now, Pool() would help us maintain a flat list of equal resources.
This would limit the scope of this utility, since we sometimes need
to keep (and serve) resources for different kinds, from a common pool.

eg. are connections to different servers, that must have a total
limit and LRU behavior.

Note: this also introduces a /limit/ functionality, as documented