date_eval: use relativedelta and fix "on 31" date for shorter months
Using relativedelta() we can have "day=31" and still return the last
day of February (28th), say. Our code with `datetime.combine()` did
raise an exception. So, asking for the last day of month would need
a hairy expression (like '... +1month -1day')

The same way, introduce a 'on last' shortcut for the 'on 31' expr,
which makes even more sense.